+++ internal/localfile/flags_unix.go
@@ -1,0 +1,10 @@
+//go:build !windows
+
+package localfile
+
+import (
+	"os"
+	"syscall"
+)
+
+const readFlags = os.O_RDONLY | syscall.O_NONBLOCK | syscall.O_NOFOLLOW