+++ internal/localfile/flags_windows.go
@@ -1,0 +1,5 @@
+package localfile
+
+import "os"
+
+const readFlags = os.O_RDONLY